+++ app/src/main/java/com/gthr/gthrcollect/model/network/firebaserealtimedb/YugiohModel.kt
@@ -3,7 +3,6 @@ package com.gthr.gthrcollect.model.network.firebaserealtimedb
 
 import com.google.firebase.database.PropertyName
 import com.gthr.gthrcollect.utils.constants.FirebaseRealtimeDatabase
-import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -79,7 +78,7 @@ data class YugiohModel(
     var koreanOG: Int? = 0,
     @get:PropertyName(FirebaseRealtimeDatabase.LOWEST_ASK_COST)
     @set:PropertyName(FirebaseRealtimeDatabase.LOWEST_ASK_COST)
-    var lowestAskCost: Int? = 0,
+    var lowestAskCost: Double? = 0.0,
     @get:PropertyName(FirebaseRealtimeDatabase.LOWEST_ASK_ID)
     @set:PropertyName(FirebaseRealtimeDatabase.LOWEST_ASK_ID)
     var lowestAskID: String? = "",
